+++ app/component-library/components/Banners/Banner/index.ts
@@ -1,0 +1,4 @@
+export { default } from './Banner';
+export { BannerAlertSeverity } from './variants/BannerAlert/BannerAlert.types';
+export { BannerTipLogoType } from './variants/BannerTip/BannerTip.types';
+export { BannerVariant } from './Banner.types';