+++ app/components/Nav/Main/MainNavigator.js
@@ -63,7 +63,7 @@ import OrderDetails from '../../UI/FiatOnRampAggregator/Views/OrderDetails';
 import TabBar from '../../../component-library/components/Navigation/TabBar';
 import BrowserUrlModal from '../../Views/BrowserUrlModal';
 import Routes from '../../../constants/navigation/Routes';
-import { trackEvent } from '../../../util/analyticsV2';
+import AnalyticsV2 from '../../../util/analyticsV2';
 import { MetaMetricsEvents } from '../../../core/Analytics';
 import { getActiveTabUrl } from '../../../util/transactions';
 import { getPermittedAccountsByHostname } from '../../../core/Permissions';
@@ -72,6 +72,9 @@ import { isEqual } from 'lodash';
 import { selectProviderConfig } from '../../../selectors/networkController';
 import { strings } from '../../../../locales/i18n';
 import isUrl from 'is-url';
+import SDKSessionsManager from '../../Views/SDKSessionsManager/SDKSessionsManager';
+import URL from 'url-parse';
+import Logger from '../../../util/Logger';
 
 const Stack = createStackNavigator();
 const Tab = createBottomTabNavigator();
@@ -162,7 +165,6 @@ const WalletTabStackFlow = () => (
     <Stack.Screen
       name="RevealPrivateCredentialView"
       component={RevealPrivateCredential}
-      options={RevealPrivateCredential.navigationOptions}
     />
   </Stack.Navigator>
 );
@@ -230,15 +232,20 @@ const HomeTabs = () => {
   const activeConnectedDapp = useSelector((state) => {
     const activeTabUrl = getActiveTabUrl(state);
     if (!isUrl(activeTabUrl)) return [];
-
-    const permissionsControllerState =
-      state.engine.backgroundState.PermissionController;
-    const hostname = new URL(activeTabUrl).hostname;
-    const permittedAcc = getPermittedAccountsByHostname(
-      permissionsControllerState,
-      hostname,
-    );
-    return permittedAcc;
+    try {
+      const permissionsControllerState =
+        state.engine.backgroundState.PermissionController;
+      const hostname = new URL(activeTabUrl).hostname;
+      const permittedAcc = getPermittedAccountsByHostname(
+        permissionsControllerState,
+        hostname,
+      );
+      return permittedAcc;
+    } catch (error) {
+      Logger.error(error, {
+        message: 'ParseUrl::MainNavigator error while parsing URL',
+      });
+    }
   }, isEqual);
 
   const options = {
@@ -246,7 +253,7 @@ const HomeTabs = () => {
       tabBarIconKey: TabBarIconKey.Wallet,
       tabBarLabel: strings('drawer.wallet'),
       callback: () => {
-        trackEvent(MetaMetricsEvents.WALLET_OPENED, {
+        AnalyticsV2.trackEvent(MetaMetricsEvents.WALLET_OPENED, {
           number_of_accounts: accountsLength,
           chain_id: chainId,
         });
@@ -257,7 +264,7 @@ const HomeTabs = () => {
       tabBarIconKey: TabBarIconKey.Browser,
       tabBarLabel: strings('drawer.browser'),
       callback: () => {
-        trackEvent(MetaMetricsEvents.BROWSER_OPENED, {
+        AnalyticsV2.trackEvent(MetaMetricsEvents.BROWSER_OPENED, {
           number_of_accounts: accountsLength,
           chain_id: chainId,
           source: 'Navigation Tab',
@@ -346,6 +353,7 @@ const SettingsFlow = () => (
       component={AdvancedSettings}
       options={AdvancedSettings.navigationOptions}
     />
+    <Stack.Screen name="SDKSessionsManager" component={SDKSessionsManager} />
     <Stack.Screen
       name="SecuritySettings"
       component={SecuritySettings}
@@ -388,7 +396,6 @@ const SettingsFlow = () => (
     <Stack.Screen
       name="RevealPrivateCredentialView"
       component={RevealPrivateCredential}
-      options={RevealPrivateCredential.navigationOptions}
     />
     <Stack.Screen
       name="WalletConnectSessionsView"