+++ app/components/Nav/Main/RootRPCMethodsUI.js
@@ -1,23 +1,15 @@
 import React, { useState, useEffect, useCallback } from 'react';
+
 import { StyleSheet, Alert, InteractionManager } from 'react-native';
 import PropTypes from 'prop-types';
 import { connect, useSelector } from 'react-redux';
 import { ethers } from 'ethers';
 import abi from 'human-standard-token-abi';
 import { ethErrors } from 'eth-json-rpc-errors';
-import Modal from 'react-native-modal';
-import { BN } from 'ethereumjs-util';
-import { swapsUtils } from '@metamask/swaps-controller';
-import { query } from '@metamask/controller-utils';
-import BigNumber from 'bignumber.js';
-import Engine from '../../../core/Engine';
-import {
-  trackEvent,
-  trackLegacyEvent,
-  trackLegacyAnonymousEvent,
-} from '../../../util/analyticsV2';
+
 import Approval from '../../Views/Approval';
 import NotificationManager from '../../../core/NotificationManager';
+import Engine from '../../../core/Engine';
 import { strings } from '../../../../locales/i18n';
 import { hexToBN, fromWei, isZeroValue } from '../../../util/number';
 import {
@@ -26,6 +18,7 @@ import {
 } from '../../../actions/transaction';
 import PersonalSign from '../../UI/PersonalSign';
 import TypedSign from '../../UI/TypedSign';
+import Modal from 'react-native-modal';
 import WalletConnect from '../../../core/WalletConnect';
 import {
   getMethodData,
@@ -37,6 +30,7 @@ import {
   getTokenValueParamAsHex,
   isSwapTransaction,
 } from '../../../util/transactions';
+import { BN } from 'ethereumjs-util';
 import Logger from '../../../util/Logger';
 import MessageSign from '../../UI/MessageSign';
 import Approve from '../../Views/ApproveView/Approve';
@@ -50,11 +44,17 @@ import {
   toggleDappTransactionModal,
   toggleApproveModal,
 } from '../../../actions/modals';
+import { swapsUtils } from '@metamask/swaps-controller';
+import { query } from '@metamask/controller-utils';
+import Analytics from '../../../core/Analytics/Analytics';
+import BigNumber from 'bignumber.js';
 import { getTokenList } from '../../../reducers/tokens';
 import { toLowerCaseEquals } from '../../../util/general';
 import { ApprovalTypes } from '../../../core/RPCMethods/RPCMethodMiddleware';
 import { KEYSTONE_TX_CANCELED } from '../../../constants/error';
 import { MetaMetricsEvents } from '../../../core/Analytics';
+import AnalyticsV2 from '../../../util/analyticsV2';
+
 import { useTheme } from '../../../util/theme';
 import withQRHardwareAwareness from '../../UI/QRHardware/withQRHardwareAwareness';
 import QRSigningModal from '../../UI/QRHardware/QRSigningModal';
@@ -215,13 +215,13 @@ const RootRPCMethodsUI = (props) => {
             quote_vs_executionRatio: quoteVsExecutionRatio,
             token_to_amount_received: tokenToAmountReceived.toString(),
           };
-          trackLegacyEvent(event, {});
-          trackLegacyAnonymousEvent(event, parameters);
+          Analytics.trackEventWithParameters(event, {});
+          Analytics.trackEventWithParameters(event, parameters, true);
         });
       } catch (e) {
         Logger.error(e, MetaMetricsEvents.SWAP_TRACKING_FAILED);
         InteractionManager.runAfterInteractions(() => {
-          trackLegacyEvent(MetaMetricsEvents.SWAP_TRACKING_FAILED, {
+          Analytics.trackEvent(MetaMetricsEvents.SWAP_TRACKING_FAILED, {
             error: e,
           });
         });
@@ -269,7 +269,9 @@ const RootRPCMethodsUI = (props) => {
           );
           Logger.error(error, 'error while trying to send transaction (Main)');
         } else {
-          trackEvent(MetaMetricsEvents.QR_HARDWARE_TRANSACTION_CANCELED);
+          AnalyticsV2.trackEvent(
+            MetaMetricsEvents.QR_HARDWARE_TRANSACTION_CANCELED,
+          );
         }
       }
     },
@@ -528,7 +530,11 @@ const RootRPCMethodsUI = (props) => {
   // Accept pending approval using MetaMask SDK.
   const acceptPendingApproval = (id, requestData) => {
     const { ApprovalController } = Engine.context;
-    ApprovalController.accept(id, requestData);
+    try {
+      ApprovalController.accept(id, requestData);
+    } catch (err) {
+      // Ignore err if request already approved or doesn't exists.
+    }
   };
 
   const onAddCustomNetworkReject = () => {
@@ -619,7 +625,9 @@ const RootRPCMethodsUI = (props) => {
    * When user clicks on approve to connect with a dapp using the MetaMask SDK.
    */
   const onAccountsConfirm = () => {
-    acceptPendingApproval(hostToApprove.id, hostToApprove.requestData);
+    if (hostToApprove) {
+      acceptPendingApproval(hostToApprove.id, hostToApprove.requestData);
+    }
     setShowPendingApproval(false);
   };
 
@@ -651,6 +659,7 @@ const RootRPCMethodsUI = (props) => {
       <AccountApproval
         onCancel={onAccountsReject}
         onConfirm={onAccountsConfirm}
+        navigation={props.navigation}
         currentPageInformation={currentPageMeta}
       />
     </Modal>
@@ -779,7 +788,7 @@ const RootRPCMethodsUI = (props) => {
           // eslint-disable-next-line no-case-declarations
           const totalAccounts = props.accountsLength;
 
-          trackEvent(MetaMetricsEvents.CONNECT_REQUEST_STARTED, {
+          AnalyticsV2.trackEvent(MetaMetricsEvents.CONNECT_REQUEST_STARTED, {
             number_of_accounts: totalAccounts,
             source: 'PERMISSION SYSTEM',
           });