+++ app/components/Nav/Main/index.js
@@ -67,6 +67,7 @@ import {
 } from '../../../component-library/components/Toast';
 import { useEnableAutomaticSecurityChecks } from '../../hooks/EnableAutomaticSecurityChecks';
 import { useMinimumVersions } from '../../hooks/MinimumVersions';
+import navigateTermsOfUse from '../../../util/termsOfUse/termsOfUse';
 import {
   selectProviderConfig,
   selectProviderType,
@@ -321,6 +322,16 @@ const Main = (props) => {
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+
+  const termsOfUse = useCallback(async () => {
+    if (props.navigation) {
+      await navigateTermsOfUse(props.navigation.navigate);
+    }
+  }, [props.navigation]);
+
+  useEffect(() => {
+    termsOfUse();
+  }, [termsOfUse]);
 
   const openDeprecatedNetworksArticle = () => {
     Linking.openURL(MM_DEPRECATED_NETWORKS);