+++ app/components/Views/Settings/SecuritySettings/Sections/index.ts
@@ -1,6 +1,6 @@
 import AutomaticSecurityChecks from './AutomaticSecurityChecks';
 import ClearCookiesSection from './ClearCookiesSection';
-import DeleteMetaMetricsData from './DeleteMetaMetricsData/DeleteMetaMetricsData';
+import DeleteMetaMetricsData from './DeleteMetaMetricsData';
 import DeleteWalletData from './DeleteWalletData';
 import RememberMeOptionSection from './RememberMeOptionSection';
 import ProtectYourWallet from './ProtectYourWallet/ProtectYourWallet';