+++ wdio/screen-objects/testIDs/Screens/WalletView.testIds.js
@@ -21,3 +21,7 @@ export const NOTIFICATION_REMIND_ME_LATER_BUTTON_ID =
   'notification-remind-later-button';
 
 export const SECURE_WALLET_BACKUP_ALERT_MODAL = 'backup-alert';
+
+export const getAssetTestId = (token) => {
+  return `asset-${token}`;
+};